break url into its own component
